feat: cleanup AWS detector logic
