+++ pkg/common/utils.go
@@ -3,9 +3,6 @@ package common
 import (
 	"bufio"
 	"crypto/rand"
-	"encoding/base32"
-	"encoding/binary"
-	"fmt"
 	"io"
 	"math/big"
 	"strings"
@@ -53,33 +50,6 @@ func RandomID(length int) string {
 	}
 
 	return string(b)
-}
-
-func GetAccountNumFromAWSID(AWSID string) (string, error) {
-	// Function to get the account number from an AWS ID (no verification required)
-	// Source: https://medium.com/@TalBeerySec/a-short-note-on-aws-key-id-f88cc4317489
-	if len(AWSID) < 4 {
-		return "", fmt.Errorf("AWSID is too short")
-	}
-	if AWSID[4] == 'I' || AWSID[4] == 'J' {
-		return "", fmt.Errorf("Can't get account number from AKIAJ/ASIAJ or AKIAI/ASIAI keys")
-	}
-	trimmedAWSID := AWSID[4:]
-	decodedBytes, err := base32.StdEncoding.WithPadding(base32.NoPadding).DecodeString(strings.ToUpper(trimmedAWSID))
-	if err != nil {
-		return "", err
-	}
-
-	if len(decodedBytes) < 6 {
-		return "", fmt.Errorf("Decoded AWSID is too short")
-	}
-
-	data := make([]byte, 8)
-	copy(data[2:], decodedBytes[0:6])
-	z := binary.BigEndian.Uint64(data)
-	const mask uint64 = 0x7fffffffff80
-	accountNum := (z & mask) >> 7
-	return fmt.Sprintf("%012d", accountNum), nil
 }
 
 // SliceContainsString searches a slice to determine if it contains a specified string.
