+++ pkg/detectors/aws/access_keys/accesskey_test.go
@@ -1,4 +1,4 @@
-package aws
+package access_keys
 
 import (
 	"context"
